+++ contrib/opencensus-ext-pyramid/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 0.7.2
+Released 2021-01-14
+- Change blacklist to excludelist
+([#977](https://github.com/census-instrumentation/opencensus-python/pull/977))
+
 ## 0.7.1
 Released 2019-08-26